Type the BattleEngine leaf utilities and value objects
Start driving the vendored OPBE battle engine toward a clean baseline,
bottom-up. Type the self-contained leaves: Number (int|float value
object), Gauss (float random helpers), Events, and the Defense/Ship
repair-probability accessors, plus IterableUtil's key() and Iterator
generics.

Typing Gauss::getNextMsBetween's parameters also resolved two Fire
baseline entries; getShotsFromOneAttackerShipOfType now casts to int to
honour its own @return int contract and keep the float maths from
cascading to its callers.

IterableUtil::$array stays untyped for now — it is the shared collection
base and typing it needs its five subclasses (Fleet/Player/PlayerGroup/…)
done together, which is a later dedicated pass. Nets 12 PHPStan baseline
entries removed (681); 563 tests still green.

+++ legacy/app/Libraries/BattleEngine/CombatObject/Fire.php
@@ -175,9 +175,9 @@ class Fire
             log_var('$max', $max);
             log_var('$min', $min);
             log_var('$mean', $meanShots);
-            return Gauss::getNextMsBetween($meanShots, GeometricDistribution::getStandardDeviationFromProbability(1 - $p), $min, $max);
+            return (int) Gauss::getNextMsBetween($meanShots, GeometricDistribution::getStandardDeviationFromProbability(1 - $p), $min, $max);
         }
-        return $meanShots;
+        return (int) $meanShots;
     }
 
     /**
